requirements.txt
rich
python-dotenv
beautifulsoup4
langchain
linkedin-python